changing type for height and weight in user controller and calorie util.

+++ bluewallFrontEndServices/bluewallFrontEndServices/src/main/java/com/bluewall/feservices/controller/RegistrationController.java
@@ -55,7 +55,7 @@ public class RegistrationController {
 		
 		UserDailyNutritionPlan dailyPlan = new UserDailyNutritionPlan();
 		double dailyCalorieRequirement;
-		double targetWeight;
+		float targetWeight;
 		
 		if (profile.getTargetWeight() != 0.0)
 			targetWeight = profile.getTargetWeight();
