+++ bluewallUtility/bluewallUtility/src/main/java/com/bluewall/util/utility/CalorieUtil.java
@@ -16,12 +16,12 @@ public class CalorieUtil {
      * @param activityLevel
      * @return calories needed everyday.
      */
-    public static double calculateDailyCalorieNeeds(double weightInKg, double heightInCm,
+    public static double calculateDailyCalorieNeeds(float weightInKg, float heightInCm,
                                              int age, Gender gender, ActivityLevel activityLevel) {
         return calculateBMR(weightInKg, heightInCm, age, gender) * activityLevel.getActivityFactor();
     }
 
-    private static double calculateBMR(double weightInKg, double heightInCm, int age, Gender gender) {
+    private static double calculateBMR(float weightInKg, float heightInCm, int age, Gender gender) {
         return 9.99 * weightInKg + 6.25 * heightInCm - 4.92 * age + gender.getBRMConstant();
     }
 }
